+++ scripts/phpcs.sh
@@ -1,7 +1,8 @@
 #!/usr/bin/env bash
 
+PHP_VERSION=8.2
 TOOL_DIR=.Build/tools/phpcs
 TOOL_PACKAGE="friendsofphp/php-cs-fixer"
 TOOL_COMMAND="php-cs-fixer fix -v --diff"
 
-source scripts/runphptool.sh
+source Scripts/runphptool.sh